Divide files to InputView, CommandLineReader and FileReader
Modify absolute path to relative path in FileReader. The relative path indicates user's Document directory
